Styled Components incorporated into a Lottery game
An attempt to incorporate Styled Components into a simple Lottery style game. My hope was to have all logic and styles on as few pages and imports as possible. Really I just didnt want to write and save and import a styles.css ... This was a lot more work than just writing the file as styles.css.